Added Comments and liking/disliking
added a way to post a comment to the database

added a way to add likes and dislikes to the comments

added a way to remove likes and dislikes to comments

added error handling when removing a like/dislike will make it negative.

This allows frontend to call and display comments as well as likes and dislikes from the database.